master

+++ commands/Fun/fun_say.js
@@ -2,25 +2,26 @@ const Discord = require("discord.js");
 
 module.exports.run = async (bot, message, args) => {
 
+  if (!message.member.hasPermission("SEND_MESSAGES")) return message.channel.send("No");
+  var smthng2say = args.join(" ");
+  if (!smthng2say) return;
   message.delete();
 
-  if (!message.member.hasPermission("SEND_MESSAGES")) return message.channel.send("No");
-  if (args.length < 0) {
-    return;
-  }
-
   const RandomColour = ("000000" + Math.random().toString(16).slice(2, 8).toUpperCase()).slice(-6);
-  if (message.author) {
     if (args[0].toLowerCase() === "embed") {
-      const embed = new Discord.RichEmbed()
+      if (args.slice(1).length <= 2000) {
+
+      const embed = new Discord.MessageEmbed()
           .setDescription(args.slice(1).join(" "))
           .setColor(RandomColour);
   
         message.channel.send(embed);
+      } else {
+        message.channel.send(smthng2say);
       }
     } else {
-      message.channel.send(args.join(" "));
-    }
+      message.channel.send(smthng2say);
+  }
 
 }
